+++ README.md
@@ -1,6 +1,6 @@
 # КомКом React Shop
 
-This repository contains a minimal React application for a demo store named **КомКом**. It lists 20 hard-coded television products with images.
+This repository contains a minimal React application for a demo store named **КомКом**. It lists 20 hard-coded television products with images. An additional page at `/phones` displays 20 phone items.
 
 ## Development
